+++ internal/manager/setup.go
@@ -342,7 +342,6 @@ func (c *Config) adminAPIClients(
 	addresses := c.KongAdminURLs
 	clients := make([]*adminapi.Client, 0, len(addresses))
 	for _, address := range addresses {
-		// REVIEW: create clients concurrently here?
 		err := retry.Do(
 			func() error {
 				cl, err := adminapi.NewKongClientForWorkspace(ctx, address, c.KongWorkspace, c.KongAdminAPIConfig, c.KongAdminToken)
@@ -352,7 +351,6 @@ func (c *Config) adminAPIClients(
 				clients = append(clients, cl)
 				return nil
 			},
-			// REVIEW: reuse the `--kong-admin-init-retries` and `--kong-admin-init-retry-delay` here or add new flags?
 			retry.Attempts(c.KongAdminInitializationRetries),
 			retry.Delay(c.KongAdminInitializationRetryDelay),
 		)
@@ -434,7 +432,6 @@ func AdminAPIClientFromServiceDiscovery(
 
 	clients := make([]*adminapi.Client, 0, len(adminAPIs))
 	for _, adminAPI := range adminAPIs {
-		// REVIEW: Should we create clients concurrently here?
 		cl, err := factory.CreateAdminAPIClient(ctx, adminAPI)
 		if err != nil {
 			return nil, err