+++ CHANGELOG.md
@@ -128,7 +128,7 @@ Adding a new version? You'll need three changes:
 - Corrected the resource status handling logic for the fallbackConfiguration feature.
   [#7306](https://github.com/Kong/kubernetes-ingress-controller/pull/7306)
 - Fixed an issue in translation to expression routes that match rules other than
-  path (hostname, headers, ...) disappear when the `HTTPRoute` rule contains a
+  path (hostname, headers, ...) would disappear when the `HTTPRoute` rule contains a
   `URLRewrite` filter with the `ReplaceMatchPrefix` path modifier.
   [#7269](https://github.com/Kong/kubernetes-ingress-controller/pull/7269)